Separate L1 block sync
