+++ crates/prover/src/lib.rs
@@ -6,6 +6,7 @@ use sov_modules_stf_blueprint::StfBlueprint;
 use tokio::sync::oneshot;
 use tracing::instrument;
 
+mod da_block_handler;
 pub mod prover_service;
 mod runner;
 pub use runner::*;
