+++ crates/sovereign-sdk/module-system/sov-modules-rollup-blueprint/src/lib.rs
@@ -57,11 +57,13 @@ pub trait RollupBlueprint: Sized + Send + Sync {
 
     /// Prover service.
     type ProverService: ProverService<
-        Self::Vm,
-        StateRoot = <<Self::NativeContext as Spec>::Storage as Storage>::Root,
-        Witness = <<Self::NativeContext as Spec>::Storage as Storage>::Witness,
-        DaService = Self::DaService,
-    >;
+            Self::Vm,
+            StateRoot = <<Self::NativeContext as Spec>::Storage as Storage>::Root,
+            Witness = <<Self::NativeContext as Spec>::Storage as Storage>::Witness,
+            DaService = Self::DaService,
+        > + Send
+        + Sync
+        + 'static;
 
     /// Creates a new instance of the blueprint.
     fn new() -> Self;
